requirements.txt
Flask
gunicorn
numpy
tensorflow-cpu
opencv-python-headless
